refactor: extract version + compare to shared
