+++ blocks/shared/utils.js
@@ -302,6 +302,13 @@ export const getAemSiteToken = (() => {
   };
 })();
 
+export function formatDate(timestamp) {
+  const rawDate = timestamp ? new Date(timestamp) : new Date();
+  const date = rawDate.toLocaleDateString([], { year: 'numeric', month: 'short', day: 'numeric' });
+  const time = rawDate.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' });
+  return { date, time };
+}
+
 export function delay(ms) {
   return new Promise((res) => { setTimeout(res, ms); });
 }
