+++ blocks/shared/version/helpers.js
@@ -1,9 +1,4 @@
-export function formatDate(timestamp) {
-  const rawDate = timestamp ? new Date(timestamp) : new Date();
-  const date = rawDate.toLocaleDateString([], { year: 'numeric', month: 'short', day: 'numeric' });
-  const time = rawDate.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' });
-  return { date, time };
-}
+import { formatDate } from '../utils.js';
 
 export function formatVersions(json) {
   // Sort by timestamp epoch
